defs are fixed
There was a wrong condition for the case when the actual tally is
computed. 1)A is allowed to not post computed ballots —> lists U_i^0
and U_I^1 and options that were post are not the same.
2) A is allowed to use ANY credentials or ballot it wants —> the posted
by A ballots spited into two groups: the right credentials/ballot
were/was used and the left ones. Therefore there is a bit more
complicated condition for when C may remove its ballots and compute the
actual tally without leaking a.

P.S. Pictures are not fixed yet